requirements.txt
streamlit==1.30.0
opencv-python==4.8.1.78
ultralytics==8.3.69
torch==2.2.2
torchvision==0.17.2
p==1.25.2
Pillow==10.2.0
numpy==1.25.2